refactor to simple example of lib usage
