+++ golang/go.mod
@@ -1,0 +1,7 @@
+module example
+
+go 1.12
+
+replace github.com/exmo-dev/exmo_api_lib/tree/master/golang/lib => ./lib/
+
+require github.com/exmo-dev/exmo_api_lib/tree/master/golang/lib v0.0.0-00010101000000-000000000000
